pga/pga_scrapers_v1/pgaTourCompletedTournaments/requirements.txt
Scrapy==2.12.0
python-dotenv
fastapi
uvicorn
psycopg2-binary==2.9.9
crochet==2.1.1
twisted==22.10.0
scrapy-zyte-smartproxy==2.4.1